refactor: reorganize project into package and scripts
