+++ fbfh_trade/__init__.py
@@ -1,0 +1,7 @@
+"""Core package for the fbfh-trade tools.
+
+All helper modules are consolidated here after the repository was
+restructured into a package-oriented layout.
+"""
+
+__all__ = []
