+++ fbfh_trade/company/__init__.py
@@ -1,0 +1,3 @@
+"""Company-related helpers for fbfh-trade."""
+
+__all__ = []
